+++ metatomic-torch/CHANGELOG.md
@@ -6,6 +6,16 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 
 ## [Unreleased](https://github.com/metatensor/metatomic/)
 
+### Added
+
+- Two new functions, `metatomic.torch.save_buffer` and
+  `metatomic.torch.load_system_buffer`, allow to serialize and deserialize `System`
+  objects to and from `torch.Tensor`
+
+### Changed
+
+- `metatomic.torch.save` and `metatomic.torch.load_system` are now implemented in C++
+
 <!-- Possible sections for each package:
 
 ### Added